refactor `main-page.tsx` code
