+++ src/components/reviews-list/reviews-list.tsx
@@ -3,17 +3,15 @@ import ReviewsItem from '../reviews-item/reviews-item';
 
 type ReviewsListProps = {
   currentReviews: Reviews;
-}
+};
 
-export default function ReviewsList({currentReviews}: ReviewsListProps) {
+export default function ReviewsList({ currentReviews }: ReviewsListProps) {
   return (
     <ul className="reviews__list">
-      {
-        currentReviews.map((review, i) => (
-          // eslint-disable-next-line react/no-array-index-key
-          <ReviewsItem review={review} key={i} />
-        ))
-      }
+      {currentReviews.map((review, i) => (
+        // eslint-disable-next-line react/no-array-index-key
+        <ReviewsItem review={review} key={i} />
+      ))}
     </ul>
   );
 }
