+++ src/components/sorting-form/data.ts
@@ -3,7 +3,7 @@ export const SORT_OPTIONS = [
   'Price: low to high',
   'Price: high to low',
   'Top rated first',
-] satisfies Record<SortOption, string>;
+] as const;
 
 export const enum SortOption {
   Popular,
